Rename DisplayResults to Display/Results
Move libraries/classes/DisplayResults.php to libraries/classes/Display/Results.php
